Removed akka from the project for testing.
